refactor: Moved logic from sheet-viewmodel-collection to viewmodel-collection and removed deprecated class.
fix: viewmodel-collection: Disposing of view models now properly calls their dispose method.

* This removes an unnecessary layer of indirection, as the view model collection itself doesn't actually need to care about its parent. It is the parent's responsibility to dispose of its view models, when appropriate.
